Reorganized keyboard files; added the touch mouse messages to the RA localization files; use the RA notification system
