+++ msg_hash.h
@@ -512,6 +512,9 @@ enum msg_hash_enums
    MSG_FAILED_TO_ENTER_GAMEMODE,
    MSG_FAILED_TO_ENTER_GAMEMODE_LINUX,
 
+   MSG_IOS_TOUCH_MOUSE_ENABLED,
+   MSG_IOS_TOUCH_MOUSE_DISABLED,
+
    MENU_LABEL(MENU_XMB_ANIMATION_HORIZONTAL_HIGHLIGHT),
    MENU_LABEL(MENU_XMB_ANIMATION_MOVE_UP_DOWN),
    MENU_LABEL(MENU_XMB_ANIMATION_OPENING_MAIN_MENU),
